requirements.txt
langchain
langchain_community
langchain_core
pandas
python-dotenv
SQLAlchemy
langchain-openai
pymilvus